Get double click working and rearrange how the marker is managed
