+++ src/lib/draw/index.ts
@@ -3,7 +3,6 @@ export { default as ImageLayer } from "./image/ImageLayer.svelte";
 export { default as HoverLayer } from "./HoverLayer.svelte";
 export { default as InterventionLayer } from "./InterventionLayer.svelte";
 export { default as NewFeatureForm } from "./NewFeatureForm.svelte";
-export { default as PointToolLayer } from "./point/PointToolLayer.svelte";
 export { default as PolygonToolLayer } from "./polygon/PolygonToolLayer.svelte";
 export { default as RouteSnapperLayer } from "./route/RouteSnapperLayer.svelte";
 export { default as StreetViewMode } from "./StreetViewMode.svelte";
